Agente_IA/requirements.txt
pandas
openai
streamlit
mysql-connector-python
faker